public-api: using /v1 instead of /public
