+++ packages/api-server/src/cmd/gen-docs.ts
@@ -38,7 +38,7 @@ const app: FastifyPluginAsync<AppOptions, RawServerDefault, JsonSchemaToTsProvid
         version: '0.0.1'
       },
       servers: [{
-        url: `${app.config.API_BASE_URL || 'https://api.ossinsight.io'}/public`
+        url: `${app.config.API_BASE_URL || 'https://api.ossinsight.io'}/v1`
       }],
     },
     hideUntagged: true,
@@ -46,9 +46,9 @@ const app: FastifyPluginAsync<AppOptions, RawServerDefault, JsonSchemaToTsProvid
 
   // Loads all routes defined in ./routes directory.
   await app.register(AutoLoad, {
-    dir: join(__dirname, '../routes/public'),
+    dir: join(__dirname, '../routes/v1'),
     routeParams: true,
-    prefix: '/public',
+    prefix: '/v1',
     options: opts
   });
