+++ packages/api-server/src/routes/v1/repos/_owner/_repo/pull_request_creators/organizations/index.ts
@@ -2,14 +2,27 @@ import {FastifyPluginAsync} from "fastify";
 import {proxyGet, getSuccessResponse} from "../../../../../../../utils/endpoint";
 
 const schema = {
-  operationId: 'list-organizations-of-issue-creators',
-  summary: 'List organizations of stargazers',
+  operationId: 'list-organizations-of-pr-creators',
+  summary: 'List organizations of PR creators',
   method: 'GET',
-  description: 'List organizations of stargazers for the specified repository.',
-  tags: ['Issue Creators'],
+  description: `List organizations of pull request creators for the specified repository.
+
+> **Notice**:
+> In the overall data, about **5.62%** of GitHub users has valid organization information.
+
+> **Note**: 
+> By default, the API does not count users without valid organization information. 
+> If you need to count these users, you can set the \`exclude_unknown\` parameter to \`false\`.
+`,
+  tags: ['Pull Request Creators'],
   querystring: {
     type: 'object',
     properties: {
+      exclude_unknown: {
+        type: 'boolean',
+        description: 'Whether to exclude PR creators with unknown organization information',
+        default: true,
+      },
       from: {
         type: 'string',
         description: 'The start date of the range.',
@@ -46,7 +59,7 @@ const schema = {
         nullable: true
       },
       {
-        col: "issue_creators",
+        col: "pull_request_creators",
         data_type: "BIGINT",
         nullable: true
       },
@@ -60,42 +73,41 @@ const schema = {
       items: {
         type: 'object',
         properties: {
-          issue_creators: {
-            type: 'string',
-            description: 'Number of issue creators from the organization'
-          },
           org_name: {
             type: 'string',
             description: 'Name of the organization'
           },
           percentage: {
             type: 'string',
-            description: 'Percentage of issue creators from the organization'
+            description: 'Percentage of pull request creators from the organization'
           },
-        },
-        additionalProperties: true,
+          pull_request_creators: {
+            type: 'string',
+            description: 'Number of pull request creators from the organization'
+          },
+        }
       },
       example: [
         {
-          "issue_creators": "117",
           "org_name": "pingcap",
-          "percentage": "0.2833"
+          "percentage": "0.3064",
+          "pull_request_creators": "106"
         },
         {
-          "issue_creators": "9",
-          "org_name": "tencent",
-          "percentage": "0.0218"
-        },
-        {
-          "issue_creators": "7",
-          "org_name": "alibaba",
-          "percentage": "0.0169"
-        },
-        {
-          "issue_creators": "6",
           "org_name": "bytedance",
-          "percentage": "0.0145"
-        }
+          "percentage": "0.0318",
+          "pull_request_creators": "11"
+        },
+        {
+          "org_name": "alibaba",
+          "percentage": "0.0260",
+          "pull_request_creators": "9"
+        },
+        {
+          "org_name": "tencent",
+          "percentage": "0.0173",
+          "pull_request_creators": "6"
+        },
       ]
     }, {
       row_count: 4
