+++ packages/api-server/src/routes/v1/repos/_owner/_repo/stargazers/countries/index.ts
@@ -2,14 +2,27 @@ import {FastifyPluginAsync} from "fastify";
 import {proxyGet, getSuccessResponse} from "../../../../../../../utils/endpoint";
 
 const schema = {
-  operationId: 'list-countries-of-issue-creators',
-  summary: 'List countries/regions of issue creators',
+  operationId: 'list-countries-of-stargazers',
+  summary: 'List countries/regions of stargazers',
   method: 'GET',
-  description: 'List countries/regions of stargazers for the specified repository.',
-  tags: ['Issue Creators'],
+  description: `List countries/regions of stargazers for the specified repository.
+  
+> **Notice**:
+> In the overall data, about **3.5%** of GitHub users provided valid country/region information.
+
+> **Note**: 
+> By default, the API does not count users without valid country/region information. 
+> If you need to count these users, you can set the \`exclude_unknown\` parameter to \`false\`.
+`,
+  tags: ['Stargazers'],
   querystring: {
     type: 'object',
     properties: {
+      exclude_unknown: {
+        type: 'boolean',
+        description: 'Whether to exclude issue creators with unknown country/region information',
+        default: true,
+      },
       from: {
         type: 'string',
         description: 'The start date of the range.',
@@ -46,15 +59,16 @@ const schema = {
         nullable: true
       },
       {
+        col: "stargazers",
+        data_type: "BIGINT",
+        nullable: true
+      },
+      {
         col: "percentage",
         data_type: "DECIMAL",
         nullable: true
       },
-      {
-        col: "issue_creators",
-        data_type: "BIGINT",
-        nullable: true
-      },
+
     ], {
       type: 'array',
       items: {
@@ -64,42 +78,41 @@ const schema = {
             type: 'string',
             description: 'Country/region code'
           },
-          issue_creators: {
+          stargazers: {
             type: 'string',
-            description: 'Number of issue creators from the country/region'
+            description: 'Number of stargazers from the country/region'
           },
           percentage: {
             type: 'string',
-            description: 'Percentage of issue creators from the country/region'
+            description: 'Percentage of stargazers from the country/region'
           },
-        },
-        additionalProperties: true,
+        }
       },
       example: [
         {
           "country_code": "CN",
-          "issue_creators": "7131",
-          "percentage": "0.8749"
+          "percentage": "0.5935",
+          "stargazers": "9189"
         },
         {
           "country_code": "US",
-          "issue_creators": "316",
-          "percentage": "0.0388"
+          "percentage": "0.0996",
+          "stargazers": "1542"
         },
         {
-          "country_code": "CA",
-          "issue_creators": "243",
-          "percentage": "0.0298"
+          "country_code": "JP",
+          "percentage": "0.0305",
+          "stargazers": "473"
         },
         {
-          "country_code": "NL",
-          "issue_creators": "223",
-          "percentage": "0.0274"
+          "country_code": "DE",
+          "percentage": "0.0267",
+          "stargazers": "413"
         },
       ]
     }, {
       row_count: 4
-    }),
+    })
   }
 };
