+++ web/docs/api/sidebar.js
@@ -1,1 +1,1 @@
-module.exports = [{ type: 'doc', id: 'api/ossinsight-public-api' }, { type: 'category', label: 'Trends', collapsed: false, items: [{ type: 'doc', id: 'api/list-trending-repos', label: 'List trending repos', className: 'api-method get' }] }, { type: 'category', label: 'Issue Creators', collapsed: false, items: [{ type: 'doc', id: 'api/list-issue-creators', label: 'List issue creators', className: 'api-method get' }, { type: 'doc', id: 'api/list-countries-of-issue-creators', label: 'List countries/regions of issue creators', className: 'api-method get' }, { type: 'doc', id: 'api/list-organizations-of-issue-creators', label: 'List organizations of stargazers', className: 'api-method get' }, { type: 'doc', id: 'api/issue-creators-history', label: 'Issue creators history', className: 'api-method get' }] }, { type: 'category', label: 'Pull Request Creators', collapsed: false, items: [{ type: 'doc', id: 'api/list-pull-request-creators', label: 'List pull request creators', className: 'api-method get' }, { type: 'doc', id: 'api/list-countries-of-pr-creators', label: 'List countries/regions of PR creators', className: 'api-method get' }, { type: 'doc', id: 'api/pull-request-creators-history', label: 'Pull request creators history', className: 'api-method get' }, { type: 'doc', id: 'api/list-organizations-of-pr-creators', label: 'List organizations of PR creators', className: 'api-method get' }] }, { type: 'category', label: 'Stargazers', collapsed: false, items: [{ type: 'doc', id: 'api/stargazers-history', label: 'Stargazers history', className: 'api-method get' }, { type: 'doc', id: 'api/list-organizations-of-stargazers', label: 'List organizations of stargazers', className: 'api-method get' }, { type: 'doc', id: 'api/list-countries-of-stargazers', label: 'List countries/regions of stargazers', className: 'api-method get' }] }];
+module.exports = [{ type: 'doc', id: 'api/ossinsight-public-api' }, { type: 'category', label: 'Trends', collapsed: false, items: [{ type: 'doc', id: 'api/list-trending-repos', label: 'List trending repos', className: 'api-method get' }] }, { type: 'category', label: 'Issue Creators', collapsed: false, items: [{ type: 'doc', id: 'api/list-issue-creators', label: 'List issue creators', className: 'api-method get' }, { type: 'doc', id: 'api/issue-creators-history', label: 'Issue creators history', className: 'api-method get' }, { type: 'doc', id: 'api/list-countries-of-issue-creators', label: 'List countries/regions of issue creators', className: 'api-method get' }, { type: 'doc', id: 'api/list-organizations-of-issue-creators', label: 'List organizations of stargazers', className: 'api-method get' }] }, { type: 'category', label: 'Pull Request Creators', collapsed: false, items: [{ type: 'doc', id: 'api/list-pull-request-creators', label: 'List pull request creators', className: 'api-method get' }, { type: 'doc', id: 'api/list-countries-of-pr-creators', label: 'List countries/regions of PR creators', className: 'api-method get' }, { type: 'doc', id: 'api/list-organizations-of-pr-creators', label: 'List organizations of PR creators', className: 'api-method get' }, { type: 'doc', id: 'api/pull-request-creators-history', label: 'Pull request creators history', className: 'api-method get' }] }, { type: 'category', label: 'Stargazers', collapsed: false, items: [{ type: 'doc', id: 'api/list-organizations-of-stargazers', label: 'List organizations of stargazers', className: 'api-method get' }, { type: 'doc', id: 'api/stargazers-history', label: 'Stargazers history', className: 'api-method get' }, { type: 'doc', id: 'api/list-countries-of-stargazers', label: 'List countries/regions of stargazers', className: 'api-method get' }] }];
